Search Prosecution Cases by their reference
API:SearchProsecutionCase service acts as an orchestrator, triggering
the search into the HMCTS Common Platform API for a Prosecution Case and
recording the response result/s into Database.

The search for Prosecution Case can be done by multiple parameters, so
this first "search by Prosecution Case Reference" is an initial step
that helps us to pipe the full process from the order to search to the
storage of the results.

Other parameters for search have to come afterwards.