Refactor API endpoints and add versioning
Restructure the app so that all end points are defined in an
'/api/<internal|external>/<version>/<name>' format.

This allows for versioning to be added and also adds clarity for
endpoints with the same name but called from different places (for
example we will have two hearings endpoints - one to be called by CPUI
for searching, the other to be called by HMCTS to post updates. This
approach allows the two to be kept clearly separate.)