CACP-233 Define Unlink JSON Schema
Adapt laa_reference.json to allow for a DELETE request to unlink LAA
References.

This request includes the LAA Reference to be unlinked but no defendant
id, so the change involves a new definition resource_to_delete, which
replicates the new_resource definition with the omission of the
defendant relationship.