Add offences and other defendant data
The CPUI needs to display offence data and other defendant data.

This commit adds gender and address detail to the defendant model.

It also adds an Offence model to hold offence data returned from CP, and
adds a has_many relationship to the Defendant model.

Serializers and specs are updated appropriately.