Make CP Link calls on request from View Court Data
View Court Data provides a maat_reference and defendant_id in its
link request.

Amend the laa_references_controller so that when a link request is
received from View Court Data, the correct offence_ids for that defendant
are identified and a call is made to Common Platform to register the
maat_reference against each one.