Expose maat_reference on defendant
Replace the is_linked attribute on the defendant model with
maat_reference. This returns the maat_reference of the defendant's first
offence, unless the maat_reference begins with 'Z' - this indicates that
the case has been unlinked, and so nil is returned instead.

is_linked is no longer required as View Court Data can infer this from
the presence (or not) of a maat_reference