CACP-184 Link with no MAAT reference
Add a sequence to the database. If a request is made to link a case
without a MAAT reference, make a call to the database to get the next
sequence value, format it correctly and use it as the MAAT reference.

To allow the sequence to be correctly defined in the schema, this
necessitates changing the format of the schema dump from schema.rb
to structure.sql.