Push magistrates hearing results to SQS
Add a HearingsCreatorJob class. When a hearing is received by the
HearingsController this job will be queued, and will call the new
HearingsCreator class, which in turn will call the new
sqs::PublishMagistratesHearing class for each defendant on each
prosecution case within that hearing.

This class extracts the required data from the received hearing payload,
assembles it in the format expected by MAAT-API and pushes it to the
correct SQS queue.