Fix incorrect hearing attributes
The maatId, isActive and sessions attributes in hearing payloads pushed
to SQS for consumption by MAAT-API are incorrect.

maatId should be an integer, not a string. Add conversion and also add a
condition to filter out any dummy ids, which we do not need to receive
updates for.

isActive should be inActive and contain 'Y'/'N' not true/false.

sessions should be session (singular) and contain a single hash, not an
array of hashes.