Refactor inactive flag
The inactive flag passed to MAAT-API in mag hearing updates is currently
derived from the caseStatus received from CP.

This is incorrect as the inactive flag is used by MLRA to identify if a
case is active in the magistrates court, whereas the caseStatus in CP
will be active if the case is open in either the mags court or crown
court.

This commit amends the inactive flag so it is more accurately derived
from the court jurisdictionType instead of the caseStatus, and removes
references to caseStatus from elsewhere in the code.