Crown Court Appeals
Add functionality to process crown court appeals received from
Common Platform.

This mainly is a change to the
CrownCourtOutcomeCreator class to populate the CrownCourtOutcome part of
the payload sent to MAAT-API with appeal data instead of trial data.

This also involves amending the HearingsCreator and PublishHearing
classes to propagate relevant appeal data to the CrownCourtOutcomeCreator
class.