CACP-334 Remove references to magistrates hearings
Several classes used to publish hearings received from HMCTS refer to
magistrates courts. It was originally thought that magistrates and
crown court hearings would be substantially different and require
separate classes but it now seems that they are similar enough that they
can use the same code.

This commit removes references to magistrates in class names and
variable names and makes the code more generic.