Match format of Hearing Resulted schema
The CP format for hearing resulted includes an object of the form
```
{ hearing: {}, sharedTime: '' }
```
which was a format before, but along the course of changes,
it was changed to be just the value of `hearing`.
From the latest schemas/CP output it seems to be using the previous format
containing a `sharedTime`. An additional benefit to this is that both the
hearings resulted and the hearing notification (PUSH from HMCTS) have the same
format and that makes the consumption by CDA easier.

Additionally, in case of a hearing which isn't available, CP returns 200 with
an empty body.

Lastly, re-recorded all the VCRs from CP SIT, and tweaked the specs to use them.