Pluralize the HearingEventSerializer `type`
Currently type is not explictly set, resulting in
the type being automatically set to :hearing_event
(singular). This breaks the pattern for other serializers
that are plural.

The result was even when `hearings.hearing_events` was "included"
in a query, and hearing events returned, the consumer (VCD) was unable
to retrieve the hearing events following there patterns.

The type should, ideally, be being tested for all serializers as
it can silently break consumers when wrong.