Pluralize the ProviderSerializer `type`
Currently type is not explictly set, resulting in
the type being automatically set to :provider
(singular). This breaks the pattern for other serializers
that are plural.

The result was even when `hearings.providers` was "included"
in a query, and providers returned, the consumer (VCD) was unable
to retrieve the providers following that pattern.