Ignore dummy_reference boolean column
This is to ensure all references to the attribute have been removed and deployed,
before the migration to drop the column from the database has been deployed and run.

Using this two step approach to dropping columns ensures there is no code that
raises errors due to having a cached schema in memory at the time the schema
migration is run.

https://api.rubyonrails.org/classes/ActiveRecord/ModelSchema/ClassMethods.html#method-i-ignored_columns-3D