Use to_datetime instead of to_date for more granular time comparison
When hearings are not all in the past, or not all in the future,
the relevant hearing is the most recent one in the past.

However, when we process a defendant on the same day as their
most recent hearing, our time comparison is not granular enough
(days) to determine that a hearing that is a few hours old is indeed
in the past.

This commit makes sure we use datetime rather than date to be able
to handle this particular case.