Enforce stricter encryption rules in .gitattributes
With the current rules, we list the files we want to encrypt,
according to their exact paths. This makes it too easy to accidentally
commit unencrypted files, e.g. by simply moving a file to another folder,
or renaming it.

This PR makes sure our inclusion rules are much broader, and exclusion rules
are narrow or downright specific.