Do not post to MAAT queue if post to CP fails
Currently, CDA will publish the linking payload to the MAAT queue
even if the HTTP POST requests to Common Platform failed, e.g. due
to connectivity issues.

When this happens, CDA and MAAT get out of sync with Common Platform,
and case workers working in VCD cannot make new linking attempts.

This proposed change ensures that CDA will publish to the MAAT queue
only if the calls to Common Platform were successful. That way, we
hope to prevent CDA and MAAT from getting out of sync with CP.

Because this logic is inside a background job that implements retries
on failure (with exponential backoff), it also means that once CP
connectivity is restored, the linking jobs will eventually be processed
with no input from a caseworker.